requirements.txt
fastapi==0.115.12
langchain_core==0.3.62
langchain_openai==0.3.18
numpy==2.2.6
pandas==2.2.3
PyJWT==2.10.1
PyJWT==2.10.1
python-dotenv==1.1.0
redis==5.2.1
SQLAlchemy==2.0.41
uvicorn[standard]
python-multipart
openpyxl

